fix: DBSelectLatestDeploymentAttempt is only used in tests
- ... so move it to testutil
- remove unused GetQueuedVersionFromDB
